Se termina el login

+++ src/main/java/com/example/springfunko/web/funkos/controllers/FunkoWebController.java
@@ -6,11 +6,11 @@ import com.example.springfunko.rest.funkos.dto.FunkoCreateDto;
 import com.example.springfunko.rest.funkos.dto.FunkoResponseDto;
 import com.example.springfunko.rest.funkos.dto.FunkoUpdateDto;
 import com.example.springfunko.rest.funkos.services.FunkoServiceImpl;
+import com.example.springfunko.web.funkos.store.UserStore;
 import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.MessageSource;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
@@ -20,25 +20,61 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
+import java.util.Date;
 import java.util.Locale;
 import java.util.Optional;
 import java.util.stream.Stream;
 
 @Controller
-//"funkos" and ""
 @RequestMapping(path = {"/funkos", ""})
 @Slf4j
 public class FunkoWebController {
     private final FunkoServiceImpl funkoService;
     private final CategoryService categoryService;
-    private final MessageSource messageSource;
-    //user store
+    private final UserStore userSession;
+    private static final String REDIRECT_FUNKOS = "redirect:/funkos";
+    private static final String USER_NOT_LOGGED = "User not logged";
+    private static final String LOGIN = "/login";
 
     @Autowired
-    public FunkoWebController(FunkoServiceImpl funkoService, CategoryService categoryService, MessageSource messageSource) {
+    public FunkoWebController(FunkoServiceImpl funkoService, CategoryService categoryService, UserStore userSession) {
         this.funkoService = funkoService;
         this.categoryService = categoryService;
-        this.messageSource = messageSource;
+        this.userSession = userSession;
+    }
+
+    @GetMapping("/login")
+    public String login(HttpSession session) {
+        log.info("Login GET");
+        if (isLoggedAndSessionIsActive(session)) {
+            log.info("User already logged");
+            return REDIRECT_FUNKOS;
+        }
+        return "login";
+    }
+
+    @PostMapping
+    public String login(@RequestParam("password") String password, HttpSession session, Model model) {
+        log.info("Login POST");
+        if ("pass".equals(password)) {
+            userSession.setLastLogin(new Date());
+            userSession.setLogged(true);
+            session.setAttribute("userSession", userSession);
+            session.setMaxInactiveInterval(1800);
+            return REDIRECT_FUNKOS;
+        } else {
+            return "login";
+        }
+    }
+
+    @GetMapping("/logout")
+    public String logout(HttpSession session) {
+        log.info("Logout GET");
+        session.invalidate();
+        return REDIRECT_FUNKOS;
     }
 
     @GetMapping(path = {"", "/", "/index", "/list"})
@@ -48,24 +84,39 @@ public class FunkoWebController {
                         @RequestParam(defaultValue = "0") int page,
                         @RequestParam(defaultValue = "10") int size,
                         @RequestParam(defaultValue = "id") String sortBy,
-                        @RequestParam(defaultValue = "asc") String direction,
-                        Locale locale
+                        @RequestParam(defaultValue = "asc") String direction
     ) {
+        if (!isLoggedAndSessionIsActive(session)) {
+            log.info(USER_NOT_LOGGED);
+            return REDIRECT_FUNKOS + LOGIN;
+        }
+
         log.info("Index GET with params page: {}, size: {}, sortBy: {}, direction: {}", page, size, sortBy, direction);
 
         Sort sort = direction.equalsIgnoreCase(Sort.Direction.ASC.name()) ? Sort.by(sortBy).ascending() : Sort.by(sortBy).descending();
         Pageable pageable = PageRequest.of(page, size, sort);
-        System.out.println(funkoService.findAll(Optional.empty(), Optional.empty(), Optional.empty(), pageable).getContent());
         var funkoPage = funkoService.findAll(search, Optional.empty(), Optional.empty(), pageable);
+
+        UserStore sessionData = (UserStore) session.getAttribute("userSession");
+        sessionData.incrementLoginCount();
+        var numVisits = sessionData.getLoginCount();
+        var lastLogin = sessionData.getLastLogin();
+        var localizedDate = getLocalizedDate(lastLogin, Locale.getDefault());
 
         model.addAttribute("search", search.orElse(""));
         model.addAttribute("funkosPage", funkoPage);
+        model.addAttribute("numVisits", numVisits);
+        model.addAttribute("lastLoginDate", localizedDate);
 
         return "index";
     }
 
     @GetMapping("/details/{id}")
     public String details(@PathVariable("id") Long id, Model model, HttpSession session) {
+        if (!isLoggedAndSessionIsActive(session)) {
+            log.info(USER_NOT_LOGGED);
+            return REDIRECT_FUNKOS + LOGIN;
+        }
         log.info("Details GET with id: {}", id);
 
         FunkoResponseDto funko = funkoService.findById(id);
@@ -76,6 +127,10 @@ public class FunkoWebController {
 
     @GetMapping("/create")
     public String createForm(Model model, HttpSession session) {
+        if (!isLoggedAndSessionIsActive(session)) {
+            log.info(USER_NOT_LOGGED);
+            return REDIRECT_FUNKOS + LOGIN;
+        }
         log.info("CREATE GET");
 
         Stream<String> categorias = categoryService.findAll(Optional.empty(), Optional.empty(), PageRequest.of(0, 1000)).get().map(Categoria::getName);
@@ -95,14 +150,17 @@ public class FunkoWebController {
         log.info("CREATE POST");
 
         var res = funkoService.save(funkoCreateDto);
-        System.out.println(res);
-        return "redirect:/funkos";
+        log.info("CREATE POST with res: {}", res);
+        return REDIRECT_FUNKOS;
     }
-
 
 
     @GetMapping("/update/{id}")
     public String updateForm(@PathVariable("id") Long id, Model model, HttpSession session) {
+        if (!isLoggedAndSessionIsActive(session)) {
+            log.info(USER_NOT_LOGGED);
+            return REDIRECT_FUNKOS + LOGIN;
+        }
         log.info("UPDATE GET with id: {}", id);
 
         Stream<String> categorias = categoryService.findAll(Optional.empty(), Optional.empty(), PageRequest.of(0, 1000)).get().map(Categoria::getName);
@@ -123,12 +181,16 @@ public class FunkoWebController {
         log.info("UPDATE POST with id: {}", id);
 
         var res = funkoService.update(funkoUpdateDto, id);
-        System.out.println(res);
-        return "redirect:/funkos";
+        log.info("UPDATE POST with res: {}", res);
+        return REDIRECT_FUNKOS;
     }
 
     @GetMapping("/update-image/{id}")
     public String updateImageForm(@PathVariable("id") Long funkoId, Model model, HttpSession session) {
+        if (!isLoggedAndSessionIsActive(session)) {
+            log.info(USER_NOT_LOGGED);
+            return REDIRECT_FUNKOS + LOGIN;
+        }
         log.info("UPDATE GET with funkoId: {}", funkoId);
 
         FunkoResponseDto funko = funkoService.findById(funkoId);
@@ -143,15 +205,32 @@ public class FunkoWebController {
 
         funkoService.updateImage(funkoId, imagen);
 
-        return "redirect:/funkos";
+        return REDIRECT_FUNKOS;
     }
 
     @GetMapping("/delete/{id}")
     public String deleteFunko(@PathVariable("id") Long id, HttpSession session) {
+        if (!isLoggedAndSessionIsActive(session)) {
+            log.info(USER_NOT_LOGGED);
+            return REDIRECT_FUNKOS + LOGIN;
+        }
         log.info("DELETE GET with id: {}", id);
 
         funkoService.deleteById(id);
 
-        return "redirect:/funkos";
+        return REDIRECT_FUNKOS;
     }
+
+    private String getLocalizedDate(Date date, Locale locale) {
+        LocalDateTime localDateTime = date.toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime();
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy HH:mm:ss").withLocale(locale);
+        return localDateTime.format(formatter);
+    }
+
+    private boolean isLoggedAndSessionIsActive(HttpSession session) {
+        UserStore sessionData = (UserStore) session.getAttribute("scopedTarget.userSession");
+        return sessionData != null && sessionData.isLogged();
+    }
+
+
 }
